Fix JRUBY-3563 - fix Socket.getnameinfo to accept getpeername output
Since the move toward compatibility of getpeername which not returns
a string encoded sock_addr, Socket.getnameinfo wasn't working anymore.
This patch adds this format to getnameinfo (BTW, this also solves one of
the net-ssh issue).
Also this patch implements the case when NI_NUMERICSERV is not given and
we have to give back the service as a string.

Signed-off-by: Brice Figureau <brice@daysofwonder.com>